Add NVC support
This adds support for NVC [1], a GPL3 VHDL compiler and simulator. NVC
supports co-simulation with VHPI. Most basic features of VHPI are
supported. However, plusargs, record types, and generics/constants are
not supported.

NVC supports partial re-analysis and dependency generation, and the
supplied Makefile supports both. For the initial build, everything is
analyzed in VHDL_LIB_ORDER. On subsequent builds, the generated
dependencies are used to ensure a proper ordering for analysis and
elaboration.

Some recipes are generated with foreach/eval since pattern rules are
not expressive enough to specify both the output file and the library
order-only dependency. An alternative approach could be to use
`.SECONDEXPANSION`.

[1] https://github.com/nickg/nvc